INS-2588: separate watermill support code
