+++ network/servicenetwork/servicenetwork.go
@@ -57,16 +57,16 @@ import (
 	"sync"
 	"time"
 
+	"github.com/ThreeDotsLabs/watermill/message"
+	"github.com/ThreeDotsLabs/watermill/message/router/middleware"
+
 	"github.com/insolar/insolar/network/gateway"
 	"github.com/insolar/insolar/network/gateway/bootstrap"
 
-	"github.com/ThreeDotsLabs/watermill/message"
-	"github.com/ThreeDotsLabs/watermill/message/router/middleware"
 	"github.com/pkg/errors"
 	"go.opencensus.io/trace"
 
 	"github.com/insolar/insolar/insolar/bus"
-	"github.com/insolar/insolar/insolar/payload"
 	"github.com/insolar/insolar/network/controller/common"
 
 	"github.com/insolar/insolar/component"
@@ -85,10 +85,6 @@ import (
 	"github.com/insolar/insolar/network/transport"
 )
 
-const deliverWatermillMsg = "ServiceNetwork.processIncoming"
-
-var ack = []byte{1}
-
 // ServiceNetwork is facade for network.
 type ServiceNetwork struct {
 	cfg configuration.Configuration
@@ -102,7 +98,7 @@ type ServiceNetwork struct {
 	NodeKeeper          network.NodeKeeper          `inject:""`
 	TerminationHandler  insolar.TerminationHandler  `inject:""`
 	Pub                 message.Publisher           `inject:""`
-	ContractRequester   insolar.ContractRequester   `inject:""`
+	ContractRequester   insolar.ContractRequester   `inject:""` // todo: remove
 	Sender              bus.Sender                  `inject:""`
 
 	// subcomponents
@@ -332,144 +328,8 @@ func (n *ServiceNetwork) phaseManagerOnPulse(ctx context.Context, newPulse insol
 	}
 }
 
-// SendMessageHandler async sends message with confirmation of delivery.
-func (n *ServiceNetwork) SendMessageHandler(msg *message.Message) ([]*message.Message, error) {
-	ctx := inslogger.ContextWithTrace(context.Background(), msg.Metadata.Get(bus.MetaTraceID))
-	logger := inslogger.FromContext(ctx)
-	msgType, err := payload.UnmarshalType(msg.Payload)
-	if err != nil {
-		logger.Error("failed to extract message type")
-	}
-
-	err = n.sendMessage(ctx, msg)
-	if err != nil {
-		n.replyError(ctx, msg, err)
-		return nil, nil
-	}
-
-	logger.WithFields(map[string]interface{}{
-		"msg_type":       msgType.String(),
-		"correlation_id": middleware.MessageCorrelationID(msg),
-	}).Info("Network sent message")
-
-	return nil, nil
-}
-
-func (n *ServiceNetwork) sendMessage(ctx context.Context, msg *message.Message) error {
-	node, err := n.wrapMeta(msg)
-	if err != nil {
-		return errors.Wrap(err, "failed to send message")
-	}
-
-	// Short path when sending to self node. Skip serialization
-	origin := n.NodeKeeper.GetOrigin()
-	if node.Equal(origin.ID()) {
-		err := n.Pub.Publish(bus.TopicIncoming, msg)
-		if err != nil {
-			return errors.Wrap(err, "error while publish msg to TopicIncoming")
-		}
-		return nil
-	}
-	msgBytes, err := messageToBytes(msg)
-	if err != nil {
-		return errors.Wrap(err, "error while converting message to bytes")
-	}
-	res, err := n.RPC.SendBytes(ctx, node, deliverWatermillMsg, msgBytes)
-	if err != nil {
-		return errors.Wrap(err, "error while sending watermillMsg to controller")
-	}
-	if !bytes.Equal(res, ack) {
-		return errors.Errorf("reply is not ack: %s", res)
-	}
-	return nil
-}
-
-func (n *ServiceNetwork) replyError(ctx context.Context, msg *message.Message, repErr error) {
-	logger := inslogger.FromContext(ctx).WithFields(map[string]interface{}{
-		"correlation_id": middleware.MessageCorrelationID(msg),
-	})
-	errMsg, err := payload.NewMessage(&payload.Error{Text: repErr.Error()})
-	if err != nil {
-		logger.Error(errors.Wrapf(err, "failed to create error as reply (%s)", repErr.Error()))
-		return
-	}
-	wrapper := payload.Meta{
-		Payload: msg.Payload,
-		Sender:  n.NodeKeeper.GetOrigin().ID(),
-	}
-	buf, err := wrapper.Marshal()
-	if err != nil {
-		logger.Error(errors.Wrapf(err, "failed to wrap error message (%s)", repErr.Error()))
-		return
-	}
-	msg.Payload = buf
-	n.Sender.Reply(ctx, msg, errMsg)
-}
-
-func (n *ServiceNetwork) wrapMeta(msg *message.Message) (insolar.Reference, error) {
-	receiver := msg.Metadata.Get(bus.MetaReceiver)
-	if receiver == "" {
-		return insolar.Reference{}, errors.New("Receiver in msg.Metadata not set")
-	}
-	receiverRef, err := insolar.NewReferenceFromBase58(receiver)
-	if err != nil {
-		return insolar.Reference{}, errors.Wrap(err, "incorrect Receiver in msg.Metadata")
-	}
-
-	latestPulse, err := n.PulseAccessor.Latest(context.Background())
-	if err != nil {
-		return insolar.Reference{}, errors.Wrap(err, "failed to fetch pulse")
-	}
-	wrapper := payload.Meta{
-		Payload:  msg.Payload,
-		Receiver: *receiverRef,
-		Sender:   n.NodeKeeper.GetOrigin().ID(),
-		Pulse:    latestPulse.PulseNumber,
-	}
-	buf, err := wrapper.Marshal()
-	if err != nil {
-		return insolar.Reference{}, errors.Wrap(err, "failed to wrap message")
-	}
-	msg.Payload = buf
-
-	return *receiverRef, nil
-}
-
 func isNextPulse(currentPulse, newPulse *insolar.Pulse) bool {
 	return newPulse.PulseNumber > currentPulse.PulseNumber && newPulse.PulseNumber >= currentPulse.NextPulseNumber
-}
-
-func (n *ServiceNetwork) processIncoming(ctx context.Context, args []byte) ([]byte, error) {
-	logger := inslogger.FromContext(ctx)
-	msg, err := deserializeMessage(bytes.NewBuffer(args))
-	if err != nil {
-		err = errors.Wrap(err, "error while deserialize msg from buffer")
-		logger.Error(err)
-		return nil, err
-	}
-	logger = inslogger.FromContext(ctx)
-	if inslogger.TraceID(ctx) != msg.Metadata.Get(bus.MetaTraceID) {
-		logger.Errorf("traceID from context (%s) is different from traceID from message Metadata (%s)", inslogger.TraceID(ctx), msg.Metadata.Get(bus.MetaTraceID))
-	}
-	// TODO: check pulse here
-
-	msgType, err := payload.UnmarshalTypeFromMeta(msg.Payload)
-	if err != nil {
-		logger.Error("failed to extract message type")
-	}
-	logger.WithFields(map[string]interface{}{
-		"msg_type":       msgType.String(),
-		"correlation_id": middleware.MessageCorrelationID(msg),
-	}).Info("Network received message")
-
-	err = n.Pub.Publish(bus.TopicIncoming, msg)
-	if err != nil {
-		err = errors.Wrap(err, "error while publish msg to TopicIncoming")
-		logger.Error(err)
-		return nil, err
-	}
-
-	return ack, nil
 }
 
 func (n *ServiceNetwork) GetState() insolar.NetworkState {
